+++ R/mixer.R
@@ -945,7 +945,8 @@ mcr_summary_for_list = function(values, references, ignore_zero) {
 #' In the standard version of the chart, the grey area represents the region in which no point can be
 #'  plotted because \eqn{MCR} cannot be lower than 1. In the log version, such a region does not exist.
 #'  However, in the latter, points having \eqn{MCR} equal to 1 have an ordinate equal to `-Inf` and
-#'  therefore cannot be plotted and generate a warning message.
+#'  therefore cannot be plotted. Moreover, whatever the version, sets of values containing only values
+#'  equal to \eqn{0} (i.e., having \eqn{MCR} values equal to \eqn{0}) are not plotted.
 #' 
 #' Points that are on the boundaries of the region of the group I belong to this group.
 #'  Points that are on the boundaries of the region of the group II belong to this group.
@@ -1132,6 +1133,17 @@ mcr_chart = function(values = NULL, references = NULL,
     if (is.null(thq)) thq = top_hazard_quotient(values, references, k = 1)
   }
   
+  
+  # Remove values that cannot be used (i.e. sets of values containing only 0s)
+  to_remove = (mcr == 0)
+  hi = hi[!to_remove]
+  mcr = mcr[!to_remove]
+  thq = thq[!to_remove]
+  
+  if (length(mcr) == 0) {
+    warning("There is no points to plot because MCR values are all equal to 0.")
+    return(NULL)
+  }
   
   # Extracting names of top hazard quotients
   thq = if (is.list(thq)) sapply(unname(thq), function(v) names(v)[1]) else names(thq)
@@ -2386,14 +2398,20 @@ mcr_summary_by_class = function(values, references, classes,
 #'  function with such a matrix is slightly faster.
 #' 
 #' The charts being created with the `ggplot2` package, they can be modified or completed afterwards
-#'  using the returned object.
+#'  using the returned objects.
 #' 
 #' Color specification can be done using the R predefined color names or hexadecimal values.
 #' 
 #' In the standard version of the chart, the grey area represents the region in which no point can be
 #'  plotted because \eqn{MCR} cannot be lower than 1. In the log version, such a region does not exist.
 #'  However, in the latter, points having \eqn{MCR} equal to 1 have an ordinate equal to `-Inf` and
-#'  therefore cannot be plotted and generate a warning message.
+#'  therefore cannot be plotted. Moreover, whatever the version, sets of values containing only values
+#'  equal to \eqn{0} (i.e., having \eqn{MCR} values equal to \eqn{0}) are not plotted.
+#' 
+#' Points that are on the boundaries of the region of the group I belong to this group.
+#'  Points that are on the boundaries of the region of the group II belong to this group.
+#'  Points that are on the boundary between the regions of the groups IIIA and IIIB belong to the
+#'  group IIIB.
 #' 
 #' \loadmathjax
 #' @template formula_miat_groups